otros cambios con respecto a eso

+++ RPGTaining/Assets/Scripts/Habilidades/Habilidades Mago/BolaDeLuz.cs
@@ -21,14 +21,14 @@ public class BolaDeLuz : Habilidad
 
 
 
-    public BolaDeLuz(string nombre, Sprite icono, string descripcion, float consumo, float coolDown) : base(nombre, icono, descripcion, consumo, coolDown)
+    public BolaDeLuz(string nombre, Sprite icono, string descripcion, int consumo, float coolDown) : base(nombre, icono, descripcion, consumo, coolDown)
     {
     }
 
-    public BolaDeLuz(string nombre) : base(nombre, null, "Descripcion de habilidad genérica", 10f, 5f)
+    public BolaDeLuz(string nombre) : base(nombre, null, "Descripcion de habilidad genérica", 10, 5f)
     {
     }
-    public BolaDeLuz() : base("Bola de Luz", null, "Descripción Bola de Luz", 10f, 3f){
+    public BolaDeLuz() : base("Bola de Luz", null, "Descripción Bola de Luz", 10, 3f){
     }
 
     private IEnumerator Cargar(){
